+++ traffic_ops/client/servercheckextensions.go
@@ -21,12 +21,12 @@ import (
 	"github.com/apache/trafficcontrol/lib/go-tc"
 )
 
-const API_TO_EXTENSION = apiBase + "/to_extensions"
+const API_TO_EXTENSION = apiBase + "/servercheck/extensions"
 
-// CreateTOExtension creates a to_extension
-func (to *Session) CreateTOExtension(toExtension tc.TOExtensionNullable) (tc.Alerts, ReqInf, error) {
+// CreateServerCheckExtension creates a servercheck extension
+func (to *Session) CreateServerCheckExtension(ServerCheckExtension tc.ServerCheckExtensionNullable) (tc.Alerts, ReqInf, error) {
 	var remoteAddr net.Addr
-	reqBody, err := json.Marshal(toExtension)
+	reqBody, err := json.Marshal(ServerCheckExtension)
 	reqInf := ReqInf{CacheHitStatus: CacheHitStatusMiss, RemoteAddr: remoteAddr}
 	if err != nil {
 		return tc.Alerts{}, reqInf, err
@@ -41,8 +41,8 @@ func (to *Session) CreateTOExtension(toExtension tc.TOExtensionNullable) (tc.Ale
 	return alerts, reqInf, err
 }
 
-// DeleteToExtension deletes a to_extension
-func (to *Session) DeleteTOExtension(id int) (tc.Alerts, ReqInf, error) {
+// DeleteServerCheckExtension deletes a servercheck extension
+func (to *Session) DeleteServerCheckExtension(id int) (tc.Alerts, ReqInf, error) {
 	URI := fmt.Sprintf("%s/%d", API_TO_EXTENSION, id)
 	resp, remoteAddr, err := to.request(http.MethodDelete, URI, nil)
 	reqInf := ReqInf{CacheHitStatus: CacheHitStatusMiss, RemoteAddr: remoteAddr}
@@ -55,15 +55,15 @@ func (to *Session) DeleteTOExtension(id int) (tc.Alerts, ReqInf, error) {
 	return alerts, reqInf, err
 }
 
-// GetTOExtensions gets all to_extensions
-func (to *Session) GetTOExtensions() (tc.TOExtensionResponse, ReqInf, error) {
+// GetServerCheckExtensions gets all servercheck extensions
+func (to *Session) GetServerCheckExtensions() (tc.ServerCheckExtensionResponse, ReqInf, error) {
 	resp, remoteAddr, err := to.request(http.MethodGet, API_TO_EXTENSION, nil)
 	reqInf := ReqInf{CacheHitStatus: CacheHitStatusMiss, RemoteAddr: remoteAddr}
 	if err != nil {
-		return tc.TOExtensionResponse{}, reqInf, err
+		return tc.ServerCheckExtensionResponse{}, reqInf, err
 	}
 	defer resp.Body.Close()
-	var toExtResp tc.TOExtensionResponse
+	var toExtResp tc.ServerCheckExtensionResponse
 	err = json.NewDecoder(resp.Body).Decode(&toExtResp)
 	return toExtResp, reqInf, err
 }