+++ CHANGELOG.md
@@ -17,8 +17,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Traffic Ops Golang Endpoints
   - /api/2.0 for all of the most recent route versions
   - /api/1.1/cachegroupparameters/{{cachegroupID}}/{{parameterID}} `(DELETE)`
-  - /api/1.5/to_extensions/:id `(DELETE)`
-  - /api/1.5/to_extensions `(GET, POST)`
   - /api/1.5/stats_summary `(POST)`
   - /api/1.1/cdns/routing
   - /api/1.1/cachegroupparameters/ `(GET, POST)`
@@ -29,6 +27,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - /api/1.5/letsencrypt/dnsrecords `GET`
   - /api/2.0/vault/ping `GET`
   - /api/2.0/vault/bucket/:bucket/key/:key/values `GET`
+  - /api/2.0/servercheck/extensions/:id `(DELETE)`
+  - /api/2.0/servercheck/extensions `(GET, POST)`
+  - /api/2.0/plugins `(GET)`
 
 ### Changed
 - Fix to traffic_ops_ort.pl to strip specific comment lines before checking if a file has changed.  Also promoted a changed file message from DEBUG to ERROR for report mode.
@@ -76,6 +77,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - /servers/totals
   - /stats_summary/create
   - /to_extensions/:id/delete
+  - /to_extensions
   - /traffic_monitor/stats
   - /types/trimmed
   - /user/current/jobs